Renamed test package and modify order of classes
