+++ university-department/src/main/java/it/university/department/controller/AddressController.java
@@ -1,5 +1,6 @@
 package it.university.department.controller;
 
+
 public final class AddressController {
 
 }
