+++ university-department/src/test/java/it/university/department/controller/address/DeleteTest.java
@@ -1,11 +1,11 @@
-package it.university.department.controller;
+package it.university.department.controller.address;
 
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.TestMethodOrder;
 import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
 import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest @TestMethodOrder(OrderAnnotation.class) @Order(14)
-public class InsertTest {
+@SpringBootTest @TestMethodOrder(OrderAnnotation.class) @Order(10)
+public class DeleteTest {
 
 }
