+++ university-department/src/test/java/it/university/department/controller/address/SelectTest.java
@@ -1,11 +1,11 @@
-package it.university.department.controller;
+package it.university.department.controller.address;
 
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.TestMethodOrder;
 import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
 import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest @TestMethodOrder(OrderAnnotation.class) @Order(16)
-public class DeleteTest {
+@SpringBootTest @TestMethodOrder(OrderAnnotation.class) @Order(7)
+public class SelectTest {
 
 }
